src/main/python/requirements.txt
grpcio
grpcio-tools
protobuf~=4.21.12